feat: api view user post
