+++ api/main_service/MainService/Controllers/PostsController.cs
@@ -195,18 +195,6 @@ public class PostsController : ControllerBase
 	{
 		var userId = "";
 
-		// Create Post Filter
-		var postFilter = Builders<Post>.Filter.Empty;
-
-		// Is Deleted Post Filter
-		postFilter = postFilter & Builders<Post>.Filter.Eq(x => x.IsDeleted, false);
-
-		// Public Post Filter
-		postFilter = postFilter & Builders<Post>.Filter.Eq(x => x.DisplayStatus, PostStatus.PUBLIC);
-
-		// Published At Filter
-		postFilter = postFilter & Builders<Post>.Filter.Lte(x => x.PublishedAt, DateTime.Now);
-
 		// Get User Id
 		try
 		{
@@ -217,112 +205,7 @@ public class PostsController : ControllerBase
 			userId = null;
 		}
 
-		// Private Post Filter
-		if (userId != null)
-		{
-			postFilter = postFilter | Builders<Post>.Filter.Eq(x => x.DisplayStatus, PostStatus.PRIVATE) & Builders<Post>.Filter.Eq(x => x.UserId, userId) & postFilter & Builders<Post>.Filter.Eq(x => x.IsDeleted, false); ;
-		}
-
-		// Hashtags Filter
-		if (pagination.Filter.Hashtags != null)
-		{
-			List<string> hashtagsNormal = new List<string>();
-
-			foreach (var hashtag in pagination.Filter.Hashtags!)
-			{
-				hashtagsNormal.Add(HelperClass.RemoveDiacritics(hashtag).ToUpper());
-			}
-
-			postFilter = postFilter & Builders<Post>.Filter.All(x => x.HashtagsNormalized, hashtagsNormal);
-		}
-
-		// Trending Post Filter
-		if (pagination.Filter.IsTrending)
-		{
-			postFilter = postFilter & Builders<Post>.Filter.Where(x => x.UpvotedBy.Count() >= 1);
-		}
-
-		// Pagination formula
-		var skipPage = (pagination.Page - 1) * pagination.Size;
-
-		// Config to get User own the Post
-		BsonDocument lookup = new BsonDocument{
-				{ "from", "account" },
-				{ "localField", "UserId" },
-				{ "foreignField", "_id" },
-				{ "as", "Users" }
-			};
-
-		// Config to specify needed fields
-		BsonDocument project = new BsonDocument{
-				{ "_id", 1 },
-				{ "User", new BsonDocument{
-					{"$arrayElemAt",
-						new BsonArray{ "$Users", 0 }
-					},
-				}},
-				{ "Content", 1},
-				{ "Medias",  new BsonDocument{
-					{"_id" , 1},
-					{"MediaType" , 1},
-					{"Url" , 1}
-				}},
-				{ "CreatedAt", 1 },
-				{ "ModifiedAt", 1 },
-				{ "PublishedAt", 1 },
-				{ "UpvotedBy", 1 },
-				{ "DownvotedBy", 1 },
-				{ "Hashtags", 1},
-				{ "CommentIds", 1},
-			};
-
-		// Config to sort created date decrease
-		BsonDocument sort = new BsonDocument{
-				{ "PublishedAt", -1 },
-				{ "_id", 1 }
-			};
-
-		// Handle full text search if content field is null or empty
-		if (!String.IsNullOrWhiteSpace(pagination.Filter.Content))
-		{
-			// Create full text search filter
-			var searchfilter = new BsonDocument {
-						{ "index", _configuration["DbIndexs:Post"] },
-						{ "text", new BsonDocument {
-							{ "query", pagination.Filter.Content },
-							{ "path", new BsonDocument {
-								{ "wildcard", "*" }
-						}},
-					}}
-				};
-
-			(var _, var postsFromRepoFTS) = await _postRepo.FindManyAsync(
-				indexFilter: searchfilter,
-				filter: postFilter,
-				lookup: lookup,
-				project: project,
-				sort: sort,
-				limit: pagination.Size,
-				skip: skipPage);
-
-			IEnumerable<BsonDocument> stages = new List<BsonDocument>();
-
-			var totalPostFTS = await _postRepo.CountDocumentAsync(
-				filter: searchfilter,
-				stages: stages);
-
-			var postsFTS = _mapper.Map<IEnumerable<ListPostDto>>(postsFromRepoFTS);
-
-			return Ok(new PaginationResDto<ListPostDto>((Int32)totalPostFTS, postsFTS));
-		}
-
-		(var totalPost, var postsFromRepo) = await _postRepo.FindManyAsync(
-			filter: postFilter,
-			lookup: lookup,
-			project: project,
-			sort: sort,
-			limit: pagination.Size,
-			skip: skipPage);
+		(var totalPost, var postsFromRepo) = await _postLogic.ViewPost(userId!, pagination);
 
 		var posts = _mapper.Map<IEnumerable<ListPostDto>>(postsFromRepo);
 
@@ -334,6 +217,12 @@ public class PostsController : ControllerBase
 	{
 		// Create Post Filter
 		var postFilter = Builders<Post>.Filter.Empty;
+
+		// Is Deleted Post Filter
+		postFilter = postFilter & Builders<Post>.Filter.Eq(x => x.IsDeleted, false);
+
+		// Public Post Filter
+		postFilter = postFilter & Builders<Post>.Filter.Eq(x => x.DisplayStatus, PostStatus.PUBLIC);
 
 		// Published At Filter
 		postFilter = postFilter & Builders<Post>.Filter.Gte(x => x.PublishedAt, DateTime.Now.AddDays(-7));
@@ -367,6 +256,29 @@ public class PostsController : ControllerBase
 
 
 		return Ok(new HashtagsRecommend(commonHashtag, randomHashtags));
+	}
+
+	[HttpPost("account")]
+	public async Task<ActionResult<PaginationResDto<IEnumerable<PostReadDto>>>> ViewUserPost(PaginationReqDto<UserPostFilterDto> pagination)
+	{
+
+		var userId = "";
+
+		// Get User Id
+		try
+		{
+			userId = User.FindFirst(JwtTokenPayload.USER_ID)!.Value;
+		}
+		catch (Exception)
+		{
+			userId = null;
+		}
+
+		(var totalPost, var postsFromRepo) = await _postLogic.ViewUserPost(userId!, pagination);
+
+		var posts = _mapper.Map<IEnumerable<ListPostDto>>(postsFromRepo);
+
+		return Ok(new PaginationResDto<ListPostDto>((Int32)totalPost, posts));
 	}
 
 	[HttpGet("{id}")]
