+++ api/main_service/MainService/Logic/PostLogic.cs
@@ -1,8 +1,10 @@
 using MainService.Data;
 using MainService.Dtos;
 using MainService.Models;
+using MainService.Utils;
 using MongoDB.Bson;
 using MongoDB.Driver;
+using static Constants;
 
 namespace MainService.Logic;
 
@@ -13,13 +15,16 @@ public interface IPostLogic
 	Task<VoteResponeDto> PostVoteCount(string postId);
 	Task<int> SavePost(string postId, string userId);
 	Task<bool> DeletePost(string id);
+	Task<(long total, IEnumerable<Post> entities)> ViewPost(string userId, PaginationReqDto<PostFilterDto> pagination);
+	Task<(long total, IEnumerable<Post> entities)> ViewUserPost(string userId, PaginationReqDto<UserPostFilterDto> pagination);
 }
 
 public class PostLogic : IPostLogic
 {
 	private readonly IPostRepo _postRepo;
-
 	private readonly IAccountRepo _accountRepo;
+	private readonly IConfiguration _configuration;
+
 	private readonly List<BsonDocument> userLookUp = new List<BsonDocument>(){
 		new BsonDocument
 		{
@@ -37,10 +42,46 @@ public class PostLogic : IPostLogic
 		}
 	};
 
-	public PostLogic(IPostRepo postRepo, IAccountRepo accountRepo)
+	private readonly BsonDocument viewPostLookup = new BsonDocument{
+				{ "from", "account" },
+				{ "localField", "UserId" },
+				{ "foreignField", "_id" },
+				{ "as", "Users" }
+			};
+
+	private readonly BsonDocument viewPostProject = new BsonDocument{
+				{ "_id", 1 },
+				{ "User", new BsonDocument{
+					{"$arrayElemAt",
+						new BsonArray{ "$Users", 0 }
+					},
+				}},
+				{ "Content", 1},
+				{ "Medias",  new BsonDocument{
+					{"_id" , 1},
+					{"MediaType" , 1},
+					{"Url" , 1}
+				}},
+				{ "CreatedAt", 1 },
+				{ "ModifiedAt", 1 },
+				{ "PublishedAt", 1 },
+				{ "UpvotedBy", 1 },
+				{ "DownvotedBy", 1 },
+				{ "Hashtags", 1},
+				{ "CommentIds", 1},
+			};
+
+	private readonly BsonDocument viewPostSort = new BsonDocument{
+				{ "PublishedAt", -1 },
+				{ "_id", 1 }
+			};
+
+	public PostLogic(IPostRepo postRepo, IConfiguration configuration, IAccountRepo accountRepo)
 	{
 		_postRepo = postRepo;
 		_accountRepo = accountRepo;
+		_configuration = configuration;
+
 	}
 
 	public async Task<Post?> GetPostById(string id)
@@ -179,5 +220,128 @@ public class PostLogic : IPostLogic
 		var update = Builders<Post>.Update.Set(x => x.IsDeleted, true);
 		var rs = await _postRepo.UpdateOneAsync(filter: filter, update: update);
 		return rs;
+	}
+
+	public async Task<(long total, IEnumerable<Post> entities)> ViewPost(string? userId, PaginationReqDto<PostFilterDto> pagination)
+	{
+		// Create Post Filter
+		var postFilter = Builders<Post>.Filter.Empty;
+
+		// Is Deleted Post Filter
+		postFilter = postFilter & Builders<Post>.Filter.Eq(x => x.IsDeleted, false);
+
+		// Public Post Filter
+		postFilter = postFilter & Builders<Post>.Filter.Eq(x => x.DisplayStatus, PostStatus.PUBLIC);
+
+		// Published At Filter
+		postFilter = postFilter & Builders<Post>.Filter.Lte(x => x.PublishedAt, DateTime.Now);
+
+		// Private Post Filter
+		if (userId != null)
+		{
+			postFilter = postFilter | Builders<Post>.Filter.Eq(x => x.DisplayStatus, PostStatus.PRIVATE)
+									& Builders<Post>.Filter.Eq(x => x.UserId, userId)
+									& Builders<Post>.Filter.Eq(x => x.IsDeleted, false);
+		}
+
+		// Hashtags Filter
+		if (pagination.Filter.Hashtags != null)
+		{
+			List<string> hashtagsNormal = new List<string>();
+
+			foreach (var hashtag in pagination.Filter.Hashtags!)
+			{
+				hashtagsNormal.Add(HelperClass.RemoveDiacritics(hashtag).ToUpper());
+			}
+
+			postFilter = postFilter & Builders<Post>.Filter.All(x => x.HashtagsNormalized, hashtagsNormal);
+		}
+
+		// Trending Post Filter
+		if (pagination.Filter.IsTrending)
+		{
+			postFilter = postFilter & Builders<Post>.Filter.Where(x => x.UpvotedBy.Count() >= 1);
+		}
+
+		// Pagination formula
+		var skipPage = (pagination.Page - 1) * pagination.Size;
+
+		// Handle full text search if content field is null or empty
+		if (!String.IsNullOrWhiteSpace(pagination.Filter.Content))
+		{
+			// Create full text search filter
+			var searchfilter = new BsonDocument {
+						{ "index", _configuration["DbIndexs:Post"] },
+						{ "text", new BsonDocument {
+							{ "query", pagination.Filter.Content },
+							{ "path", new BsonDocument {
+								{ "wildcard", "*" }
+						}},
+					}}
+				};
+
+			(var _, var postsFromRepoFTS) = await _postRepo.FindManyAsync(
+				indexFilter: searchfilter,
+				filter: postFilter,
+				lookup: viewPostLookup,
+				project: viewPostProject,
+				sort: viewPostSort,
+				limit: pagination.Size,
+				skip: skipPage);
+
+			IEnumerable<BsonDocument> stages = new List<BsonDocument>();
+
+			var totalPostFTS = await _postRepo.CountDocumentAsync(
+				filter: searchfilter,
+				stages: stages);
+
+			return (totalPostFTS, postsFromRepoFTS);
+		}
+
+		(var totalPost, var postsFromRepo) = await _postRepo.FindManyAsync(
+			filter: postFilter,
+			lookup: viewPostLookup,
+			project: viewPostProject,
+			sort: viewPostSort,
+			limit: pagination.Size,
+			skip: (pagination.Page - 1) * pagination.Size); // Pagination formula
+
+		return (totalPost, postsFromRepo);
+	}
+
+	public async Task<(long total, IEnumerable<Post> entities)> ViewUserPost(string userId, PaginationReqDto<UserPostFilterDto> pagination)
+	{
+		// Create Post Filter
+		var postFilter = Builders<Post>.Filter.Empty;
+
+		// Is Deleted Post Filter
+		postFilter = postFilter & Builders<Post>.Filter.Eq(x => x.IsDeleted, false);
+
+		// User Post Filter
+		postFilter = postFilter & Builders<Post>.Filter.Eq(x => x.UserId, pagination.Filter.UserId);
+
+		// Public Post Filter
+		postFilter = postFilter & Builders<Post>.Filter.Eq(x => x.DisplayStatus, PostStatus.PUBLIC);
+
+		// Published At Filter
+		postFilter = postFilter & Builders<Post>.Filter.Lte(x => x.PublishedAt, DateTime.Now);
+
+		// Private Post Filter
+		if (userId == pagination.Filter.UserId)
+		{
+			postFilter = postFilter | Builders<Post>.Filter.Eq(x => x.DisplayStatus, PostStatus.PRIVATE)
+									& Builders<Post>.Filter.Eq(x => x.UserId, userId)
+									& Builders<Post>.Filter.Eq(x => x.IsDeleted, false);
+		}
+
+		(var totalPost, var postsFromRepo) = await _postRepo.FindManyAsync(
+			filter: postFilter,
+			lookup: viewPostLookup,
+			project: viewPostProject,
+			sort: viewPostSort,
+			limit: pagination.Size,
+			skip: (pagination.Page - 1) * pagination.Size); // Pagination formula
+
+		return (totalPost, postsFromRepo);
 	}
 }
