+++ client/src/components/Discover/PostCreate.tsx
@@ -296,66 +296,67 @@ const PostCreate: React.FC<PostCreateProps> = ({ loading }) => {
     };
 
     return (
-        <form
-            className="card grid grid-cols-5 sm:grid-cols-10 gap-4 min-w-[20rem] p-4"
-            onSubmit={handleCreatePost}
-        >
-            {/* Avatar */}
-            <div className="col-span-1 flex justify-center ">
-                <div
-                    className="avatar"
-                    style={{
-                        backgroundImage: `url(${authState?.avatar})`,
-                    }}
-                />
-                <div />
-            </div>
-
-            <div className="col-span-4 sm:col-span-9">
-                {publishedAt && (
-                    <div className="flex items-center">
-                        <div className="text-[color:var(--text-secondary-color)] mb-2 text-xs sm:text-base">
-                            Will publish at {publishedAt.toLocaleString()}
-                        </div>
-                        <div
-                            className="cancel-circle flex justify-center items-center cursor-pointer w-[1rem] h-[1rem] text-[0.75rem] mb-2 ml-2"
-                            onClick={() => setPublishedAt(null)}
-                        >
-                            <Icon icon="times" />
-                        </div>
-                    </div>
-                )}
-
-                {/* Select Display Status */}
-                {isOpenSelect && !publishedAt ? (
-                    <Select
-                        options={options}
-                        defaultValue={displayStatus}
-                        onChange={(val: any) => setDisplayStatus(val)}
-                        styles={customStyles}
-                        components={{
-                            IndicatorSeparator: () => null,
+        <div className="lg:pt-4">
+            <form
+                className="card grid grid-cols-5 sm:grid-cols-10 gap-4 min-w-[20rem] p-4"
+                onSubmit={handleCreatePost}
+            >
+                {/* Avatar */}
+                <div className="col-span-1 flex justify-center ">
+                    <div
+                        className="avatar"
+                        style={{
+                            backgroundImage: `url(${authState?.avatar})`,
                         }}
                     />
-                ) : (
-                    ""
-                )}
+                    <div />
+                </div>
 
-                {/* Content */}
-                <textarea
-                    className="focus:outline-none bg-[color:var(--post-bg-color)] text-[color:var(--text-primary-color)] text-xl sm:text-2xl h-12 sm:h-20 w-full border-b-2 mb-4 pt-2"
-                    placeholder="How do you feel today?"
-                    rows={3}
-                    ref={inputRef}
-                    onChange={(event: React.ChangeEvent<HTMLTextAreaElement>) =>
-                        setContent(event.currentTarget.value)
-                    }
-                    value={content}
-                    onFocus={() => setIsOpenSelect(true)}
-                />
+                <div className="col-span-4 sm:col-span-9">
+                    {publishedAt && (
+                        <div className="flex items-center">
+                            <div className="text-[color:var(--text-secondary-color)] mb-2 text-xs sm:text-base">
+                                Will publish at {publishedAt.toLocaleString()}
+                            </div>
+                            <div
+                                className="cancel-circle flex justify-center items-center cursor-pointer w-[1rem] h-[1rem] text-[0.75rem] mb-2 ml-2"
+                                onClick={() => setPublishedAt(null)}
+                            >
+                                <Icon icon="times" />
+                            </div>
+                        </div>
+                    )}
 
-                {/* Might update in the future */}
-                {/* <div className="bg-[color:var(--post-bg-color)]  text-xl sm:text-2xl h-12 sm:h-20 w-full mb-[-5rem] pt-2 whitespace-pre-wrap">
+                    {/* Select Display Status */}
+                    {isOpenSelect && !publishedAt ? (
+                        <Select
+                            options={options}
+                            defaultValue={displayStatus}
+                            onChange={(val: any) => setDisplayStatus(val)}
+                            styles={customStyles}
+                            components={{
+                                IndicatorSeparator: () => null,
+                            }}
+                        />
+                    ) : (
+                        ""
+                    )}
+
+                    {/* Content */}
+                    <textarea
+                        className="focus:outline-none bg-[color:var(--post-bg-color)] text-[color:var(--text-primary-color)] text-xl sm:text-2xl h-12 sm:h-20 w-full border-b-2 mb-4 pt-2"
+                        placeholder="How do you feel today?"
+                        rows={3}
+                        ref={inputRef}
+                        onChange={(
+                            event: React.ChangeEvent<HTMLTextAreaElement>
+                        ) => setContent(event.currentTarget.value)}
+                        value={content}
+                        onFocus={() => setIsOpenSelect(true)}
+                    />
+
+                    {/* Might update in the future */}
+                    {/* <div className="bg-[color:var(--post-bg-color)]  text-xl sm:text-2xl h-12 sm:h-20 w-full mb-[-5rem] pt-2 whitespace-pre-wrap">
                     {HASHTAG_FORMATTER(content)}
                 </div>
                 <div
@@ -368,174 +369,182 @@ const PostCreate: React.FC<PostCreateProps> = ({ loading }) => {
                     suppressContentEditableWarning
                 ></div> */}
 
-                {/* Display Files */}
-                <div>
-                    {/* Image */}
-                    {typeUpload !== "image" ? (
-                        ""
-                    ) : (
-                        <div
-                            className={
-                                uploadedImages.length > 1
-                                    ? "columns-2 mb-4"
-                                    : "columns-1 mb-4"
-                            }
-                        >
-                            {uploadedImages.map(
-                                (imageUpload: ImageUploadProps) => (
-                                    <div
-                                        key={imageUpload.id}
-                                        className="relative flex justify-end pt-2 sm:pt-4"
-                                    >
-                                        <img
-                                            src={URL.createObjectURL(
-                                                imageUpload.file
-                                            )}
-                                            className="hum-image-post"
-                                            height={1000}
-                                            width={1000}
-                                        />
+                    {/* Display Files */}
+                    <div>
+                        {/* Image */}
+                        {typeUpload !== "image" ? (
+                            ""
+                        ) : (
+                            <div
+                                className={
+                                    uploadedImages.length > 1
+                                        ? "columns-2 mb-4"
+                                        : "columns-1 mb-4"
+                                }
+                            >
+                                {uploadedImages.map(
+                                    (imageUpload: ImageUploadProps) => (
                                         <div
-                                            className="cancel-circle absolute flex justify-center items-center cursor-pointer w-[1rem] h-[1rem] sm:w-[1.5rem] sm:h-[1.5rem] mt-[-0.5rem] mr-[-0.5rem] sm:mt-[-0.75rem] sm:mr-[-0.75rem] text-[0.75rem] sm:text-[1rem]"
-                                            onClick={() =>
-                                                handleRemoveImageFile(
-                                                    imageUpload.id
-                                                )
-                                            }
+                                            key={imageUpload.id}
+                                            className="relative flex justify-end pt-2 sm:pt-4"
                                         >
-                                            <Icon icon="times" />
+                                            <img
+                                                src={URL.createObjectURL(
+                                                    imageUpload.file
+                                                )}
+                                                className="hum-image-post"
+                                                height={1000}
+                                                width={1000}
+                                            />
+                                            <div
+                                                className="cancel-circle absolute flex justify-center items-center cursor-pointer w-[1rem] h-[1rem] sm:w-[1.5rem] sm:h-[1.5rem] mt-[-0.5rem] mr-[-0.5rem] sm:mt-[-0.75rem] sm:mr-[-0.75rem] text-[0.75rem] sm:text-[1rem]"
+                                                onClick={() =>
+                                                    handleRemoveImageFile(
+                                                        imageUpload.id
+                                                    )
+                                                }
+                                            >
+                                                <Icon icon="times" />
+                                            </div>
                                         </div>
-                                    </div>
-                                )
-                            )}
-                        </div>
-                    )}
+                                    )
+                                )}
+                            </div>
+                        )}
 
-                    {/* Audio */}
-                    {typeUpload !== "audio" ? (
-                        ""
-                    ) : (
-                        <ReactPlayer
-                            url={URL.createObjectURL(uploadedAudio)}
-                            controls={true}
-                            width="100%"
-                            height={50}
-                            style={{
-                                marginBottom: "1rem",
-                            }}
-                        />
-                    )}
-
-                    {/* Video */}
-                    {typeUpload !== "video" ? (
-                        ""
-                    ) : (
-                        <div className="relative pt-[56.25%] mb-4">
+                        {/* Audio */}
+                        {typeUpload !== "audio" ? (
+                            ""
+                        ) : (
                             <ReactPlayer
-                                url={URL.createObjectURL(uploadedVideo)}
+                                url={URL.createObjectURL(uploadedAudio)}
                                 controls={true}
                                 width="100%"
-                                height="100%"
+                                height={50}
                                 style={{
-                                    position: "absolute",
-                                    top: 0,
-                                    left: 0,
+                                    marginBottom: "1rem",
                                 }}
                             />
-                        </div>
-                    )}
-                </div>
+                        )}
 
-                {/* Create Post Options */}
-                <div className=" sm:flex sm:justify-between m-0">
-                    <div className="flex items-center">
-                        {/* Select Files */}
-                        <button type="button">
-                            <label htmlFor="audio-file">
-                                <Icon className="post-icon" icon="file-audio" />
-                            </label>
-                        </button>
-                        <input
-                            id="audio-file"
-                            accept="image/jpeg,image/png,image/webp,audio/mpeg,video/mp4,video/x-msvideo,video/mpeg,video/webm"
-                            type="file"
-                            onChange={handleUploadFiles}
-                            multiple
-                            hidden
-                        />
-
-                        {/* Select Icons */}
-                        <div ref={pickerRef}>
-                            <button type="button">
-                                <Icon
-                                    onClick={() => setShowPicker(!showPicker)}
-                                    className="post-icon mx-4 sm:mx-8"
-                                    icon="face-smile"
+                        {/* Video */}
+                        {typeUpload !== "video" ? (
+                            ""
+                        ) : (
+                            <div className="relative pt-[56.25%] mb-4">
+                                <ReactPlayer
+                                    url={URL.createObjectURL(uploadedVideo)}
+                                    controls={true}
+                                    width="100%"
+                                    height="100%"
+                                    style={{
+                                        position: "absolute",
+                                        top: 0,
+                                        left: 0,
+                                    }}
                                 />
-                            </button>
-                            {showPicker && (
-                                <div className="relative">
-                                    <div className="absolute z-50 mt-4 ml-[-6rem] sm:ml-[-7rem]">
-                                        <EmojiPicker
-                                            onEmojiClick={onEmojiClick}
-                                            skinTonesDisabled
-                                            width={320}
-                                            height={400}
-                                        />
-                                    </div>
-                                </div>
-                            )}
-                        </div>
-
-                        {/* Set Publish Date */}
-                        <div ref={calendarRef}>
-                            <button type="button">
-                                <Icon
-                                    onClick={() =>
-                                        setIsOpenCalendar(!isOpenCalendar)
-                                    }
-                                    className="post-icon"
-                                    icon="calendar-days"
-                                />
-                            </button>
-                            {isOpenCalendar && (
-                                <div className="relative">
-                                    <div className="absolute z-50 w-96 mt-4 ml-[-10rem]">
-                                        <DatePicker
-                                            selected={publishedAt}
-                                            onChange={handleSelectPublishedAt}
-                                            showTimeSelect
-                                            inline
-                                            isClearable
-                                            timeFormat="HH:mm"
-                                            timeIntervals={15}
-                                            timeCaption="time"
-                                            dateFormat="MMM d, yyyy h:mm aa"
-                                            minDate={new Date()}
-                                        />
-                                    </div>
-                                </div>
-                            )}
-                        </div>
+                            </div>
+                        )}
                     </div>
 
-                    {/* Create Post Button */}
-                    <Button
-                        className="mt-4 sm:mt-0 px-10"
-                        type="submit"
-                        disabled={
-                            loading ||
-                            (content === "" &&
-                                uploadedImages.length === 0 &&
-                                uploadedAudio === null &&
-                                uploadedVideo === null)
-                        }
-                    >
-                        Post
-                    </Button>
+                    {/* Create Post Options */}
+                    <div className=" sm:flex sm:justify-between m-0">
+                        <div className="flex items-center">
+                            {/* Select Files */}
+                            <button type="button">
+                                <label htmlFor="audio-file">
+                                    <Icon
+                                        className="post-icon"
+                                        icon="file-audio"
+                                    />
+                                </label>
+                            </button>
+                            <input
+                                id="audio-file"
+                                accept="image/jpeg,image/png,image/webp,audio/mpeg,video/mp4,video/x-msvideo,video/mpeg,video/webm"
+                                type="file"
+                                onChange={handleUploadFiles}
+                                multiple
+                                hidden
+                            />
+
+                            {/* Select Icons */}
+                            <div ref={pickerRef}>
+                                <button type="button">
+                                    <Icon
+                                        onClick={() =>
+                                            setShowPicker(!showPicker)
+                                        }
+                                        className="post-icon mx-4 sm:mx-8"
+                                        icon="face-smile"
+                                    />
+                                </button>
+                                {showPicker && (
+                                    <div className="relative">
+                                        <div className="absolute z-50 mt-4 ml-[-6rem] sm:ml-[-7rem]">
+                                            <EmojiPicker
+                                                onEmojiClick={onEmojiClick}
+                                                skinTonesDisabled
+                                                width={320}
+                                                height={400}
+                                            />
+                                        </div>
+                                    </div>
+                                )}
+                            </div>
+
+                            {/* Set Publish Date */}
+                            <div ref={calendarRef}>
+                                <button type="button">
+                                    <Icon
+                                        onClick={() =>
+                                            setIsOpenCalendar(!isOpenCalendar)
+                                        }
+                                        className="post-icon"
+                                        icon="calendar-days"
+                                    />
+                                </button>
+                                {isOpenCalendar && (
+                                    <div className="relative">
+                                        <div className="absolute z-50 w-96 mt-4 ml-[-10rem]">
+                                            <DatePicker
+                                                selected={publishedAt}
+                                                onChange={
+                                                    handleSelectPublishedAt
+                                                }
+                                                showTimeSelect
+                                                inline
+                                                isClearable
+                                                timeFormat="HH:mm"
+                                                timeIntervals={15}
+                                                timeCaption="time"
+                                                dateFormat="MMM d, yyyy h:mm aa"
+                                                minDate={new Date()}
+                                            />
+                                        </div>
+                                    </div>
+                                )}
+                            </div>
+                        </div>
+
+                        {/* Create Post Button */}
+                        <Button
+                            className="mt-4 sm:mt-0 px-10"
+                            type="submit"
+                            disabled={
+                                loading ||
+                                (content === "" &&
+                                    uploadedImages.length === 0 &&
+                                    uploadedAudio === null &&
+                                    uploadedVideo === null)
+                            }
+                        >
+                            Post
+                        </Button>
+                    </div>
                 </div>
-            </div>
-        </form>
+            </form>
+        </div>
     );
 };
